chat/requirements.txt
litellm>=1.70.0
requests>=2.25.1
click>=8.2.0
python-dotenv>=0.19.0 
openai>=1.75.0
fastmcp>=2.4.0
sentence-transformers>=2.2.2